APPIS/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
motor==3.1.1
pymongo==4.3.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
pydantic[email]==1.10.4
typing-extensions>=4.8.0
